# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Comentaro API
  version: 0.2.0
  description: Contrato HTTP entre a API Laravel e a aplicação web.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
security: []
servers:
  - url: /api/v1
    description: Ambiente local
paths:
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Solicita recuperação com resposta neutra
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties: {email: {type: string, format: email}}
      responses:
        '200': {$ref: '#/components/responses/Message'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /auth/reset-password:
    post:
      operationId: resetPassword
      summary: Redefine a senha com mecanismo temporário
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResetPasswordRequest'}
      responses:
        '200': {$ref: '#/components/responses/Message'}
        '422': {$ref: '#/components/responses/ValidationError'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: comentaro-session
      description: Sessão Sanctum; mutações também exigem o token CSRF emitido por /sanctum/csrf-cookie.
  responses:
...
    Message:
      description: Operação concluída
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Message'}
    ValidationError:
      description: Dados inválidos
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    TooManyRequests:
      description: Limite de tentativas excedido
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    ResetPasswordRequest:
      type: object
      required: [token, email, password, password_confirmation]
      properties:
        token: {type: string}
        email: {type: string, format: email}
        password: {type: string, format: password, minLength: 10}
        password_confirmation: {type: string, format: password, minLength: 10}
    Message:
      type: object
      required: [message]
      properties: {message: {type: string}}
    Error:
      type: object
      required: [message]
      properties:
        message: {type: string}
        code: {type: [string, 'null']}
        errors:
          type: [object, 'null']
          additionalProperties:
            type: array
            items: {type: string}
...